Fix data type expiresIn, c_nonce_expires_in

+++ waltid-libraries/protocols/waltid-openid4vc-wallet/src/commonMain/kotlin/id/walt/wallet2/handlers/WalletIssuanceHandler.kt
@@ -146,7 +146,7 @@ data class RequestTokenRequest(
 data class RequestTokenResult(
     val accessToken: String,
     val cNonce: String?,
-    val expiresIn: Int?
+    val expiresIn: Long?
 )
 
 @Serializable
